Split walker.rs into directory with expressions, statements, types, declarations, helpers
